// src/types/news.ts
export type Article = {
    source: { id?: string | null; name: string };
    author?: string | null;
    title: string;
    description?: string | null;
    url: string;
    urlToImage?: string | null;
    publishedAt: string;
    content?: string | null;
};
export type NewsResponse = { status: "ok" | "error"; totalResults?: number; articles?: Article[] };